refactor: split the cleaning module into cleaning and loading_saving module
